Improvements to Volume page
