+++ client/src/App.css
@@ -55,6 +55,10 @@ body {
 .button--green {
 	background-color: var(--green);
 }
+.button--disabled {
+	background-color: var(--text);
+	opacity: 0.6;
+}
 
 .invisible {
 	display: none;
