+++ client/src/components/SeriesCard.css
@@ -22,6 +22,7 @@
 	overflow: hidden;
 	background-color: rgba(var(--gray));
 	border-radius: 2px;
+	margin-bottom: 0.2em;
 }
 .series-card__loader::after {
 	position: absolute;
@@ -57,7 +58,7 @@
 	width: auto;
 	min-width: 100%;
 	height: 100%;
-	transition: opacity .3s ease-in-out6666 ;
+	transition: opacity 0.3s ease-in-out ;
 	opacity: 1;
 }
 .series-card__img--loading{
